requirements.txt
boto3==1.26
earthengine-api=0.1.351
gdal=3.6.4
numpy==1.24
pandas==1.5
pyarrow=10.0.1
pyproj==3.5
qgis=3.30.1
snowflake-connector-python=3.0.3
snowflake-sqlalchemy=1.4.7
sqlalchemy=1.4
utm==0.7
